CSV form UX tweaks
- Use an AsyncButton to get a spinner on the Delete button
- Change success text from "Upload successfully completed" to "Uploaded"
